refactor(openomni): break stakes contract/reference import cycle
stakes-contract.ts and stakes-reference.ts value-imported each other; the
cycle only worked because the back references sit inside superRefine
closures while createStakesSchemas() runs at contract module init - one
refactor away from a TDZ crash. Extract the shared leaf values
(STAKES_POLICY_VERSION, STAKES_THETA, hashStakesValue) into
dependency-free stakes-constants.ts; contract re-exports them so the
public API is unchanged and only contract -> reference remains. The
StakesHashValue helper type moves with hashStakesValue and is not
re-exported from contract: nothing ever imported it, and knip flags the
would-be re-export as a new dead export.
